# frontend/src/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ApprovalDecision": {
        "enum": [
          "approve",
          "reject"
        ],
        "title": "ApprovalDecision",
        "type": "string"
      },
      "ApprovalResponse": {
        "additionalProperties": false,
        "properties": {
          "approver_user_id": {
            "format": "uuid",
            "title": "Approver User Id",
            "type": "string"
          },
          "decided_at": {
            "format": "date-time",
            "title": "Decided At",
            "type": "string"
          },
          "decision": {
            "$ref": "#/components/schemas/ApprovalDecision"
          },
          "proposal_id": {
            "format": "uuid",
            "title": "Proposal Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "proposal_id",
          "version",
          "approver_user_id",
          "decision",
          "decided_at"
        ],
        "title": "ApprovalResponse",
        "type": "object"
      },
...
      "CitationRequest": {
        "additionalProperties": false,
        "properties": {
          "bundle_id": {
            "format": "uuid",
            "title": "Bundle Id",
            "type": "string"
          },
          "bundle_item_position": {
            "minimum": 0,
            "title": "Bundle Item Position",
            "type": "integer"
          },
          "claim_anchor": {
            "minLength": 1,
            "title": "Claim Anchor",
            "type": "string"
          }
        },
        "required": [
          "bundle_id",
          "bundle_item_position",
          "claim_anchor"
        ],
        "title": "CitationRequest",
        "type": "object"
      },
...
      "CreateProposalRequest": {
        "additionalProperties": false,
        "properties": {
          "additional_approver_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Additional Approver Ids",
            "type": "array"
          },
          "bundle_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Bundle Ids",
            "type": "array"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationRequest"
            },
            "title": "Citations",
            "type": "array"
          },
          "edits": {
            "items": {
              "$ref": "#/components/schemas/DocumentEditRequest"
            },
            "minItems": 1,
            "title": "Edits",
            "type": "array"
          },
          "source_session_id": {
            "format": "uuid",
            "title": "Source Session Id",
            "type": "string"
          }
        },
        "required": [
          "source_session_id",
          "edits",
          "citations"
        ],
        "title": "CreateProposalRequest",
        "type": "object"
      },
...
      "DecideProposalRequest": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "enum": [
              "approve",
              "reject"
            ],
            "title": "Decision",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_version",
          "decision"
        ],
        "title": "DecideProposalRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MergeProposalRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "minimum": 1,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "MergeProposalRequest",
        "type": "object"
      },
      "MergeProposalResponse": {
        "additionalProperties": false,
        "properties": {
          "proposal_id": {
            "format": "uuid",
            "title": "Proposal Id",
            "type": "string"
          },
          "publication": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PublicationResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/ProposalStatus"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "proposal_id",
          "version",
          "status",
          "publication"
        ],
        "title": "MergeProposalResponse",
        "type": "object"
      },
...
      "ProposalResponse": {
        "additionalProperties": false,
        "properties": {
          "approvals": {
            "items": {
              "$ref": "#/components/schemas/ApprovalResponse"
            },
            "title": "Approvals",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "format": "uuid",
            "title": "Created By",
            "type": "string"
          },
          "current_version": {
            "$ref": "#/components/schemas/ProposalVersionResponse"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "source_session_id": {
            "format": "uuid",
            "title": "Source Session Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ProposalStatus"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "workspace_id": {
            "format": "uuid",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workspace_id",
          "document_id",
          "source_session_id",
          "created_by",
          "created_at",
          "version",
          "current_version",
          "status",
          "approvals"
        ],
        "title": "ProposalResponse",
        "type": "object"
      },
      "ProposalStatus": {
        "enum": [
          "open",
          "approved",
          "rejected",
          "stale",
          "merged"
        ],
        "title": "ProposalStatus",
        "type": "string"
      },
      "ProposalVersionResponse": {
        "additionalProperties": false,
        "properties": {
          "additional_approver_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Additional Approver Ids",
            "type": "array"
          },
          "base_revision_id": {
            "format": "uuid",
            "title": "Base Revision Id",
            "type": "string"
          },
          "bundle_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Bundle Ids",
            "type": "array"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationResponse"
            },
            "title": "Citations",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "format": "uuid",
            "title": "Created By",
            "type": "string"
          },
          "edits": {
            "items": {
              "$ref": "#/components/schemas/DocumentEditResponse"
            },
            "title": "Edits",
            "type": "array"
          },
          "proposal_id": {
            "format": "uuid",
            "title": "Proposal Id",
            "type": "string"
          },
          "required_approver_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Required Approver Ids",
            "type": "array"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "proposal_id",
          "version",
          "base_revision_id",
          "edits",
          "required_approver_ids",
          "bundle_ids",
          "created_by",
          "created_at",
          "citations",
          "additional_approver_ids"
        ],
        "title": "ProposalVersionResponse",
        "type": "object"
      },
...
      "ReviseProposalRequest": {
        "additionalProperties": false,
        "properties": {
          "additional_approver_ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Additional Approver Ids"
          },
          "bundle_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Bundle Ids",
            "type": "array"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationRequest"
            },
            "title": "Citations",
            "type": "array"
          },
          "edits": {
            "items": {
              "$ref": "#/components/schemas/DocumentEditRequest"
            },
            "minItems": 1,
            "title": "Edits",
            "type": "array"
          },
          "expected_version": {
            "minimum": 1,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_version",
          "edits",
          "bundle_ids",
          "citations"
        ],
        "title": "ReviseProposalRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkspaceResponse": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "WorkspaceResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "SOT",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/workspaces/{workspace_id}": {
      "get": {
        "operationId": "get_api_v1_workspaces__workspace_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get"
      }
    },
...
    "/api/v1/workspaces/{workspace_id}/documents/{document_id}/proposals": {
      "get": {
        "operationId": "list_document_proposals",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProposalResponse"
                  },
                  "title": "Response List Document Proposals",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Proposals For Document"
      },
      "post": {
        "operationId": "create_proposal_api_v1_workspaces__workspace_id__documents__document_id__proposals_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Proposal"
      }
    },
...
    "/api/v1/workspaces/{workspace_id}/proposals/{proposal_id}": {
      "get": {
        "operationId": "read_proposal_api_v1_workspaces__workspace_id__proposals__proposal_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Proposal"
      },
      "put": {
        "operationId": "revise_proposal_api_v1_workspaces__workspace_id__proposals__proposal_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revise Proposal"
      }
    },
    "/api/v1/workspaces/{workspace_id}/proposals/{proposal_id}/decisions": {
      "post": {
        "operationId": "decide_proposal_api_v1_workspaces__workspace_id__proposals__proposal_id__decisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Decide Proposal"
      }
    },
    "/api/v1/workspaces/{workspace_id}/proposals/{proposal_id}/merge": {
      "post": {
        "operationId": "merge_proposal_api_v1_workspaces__workspace_id__proposals__proposal_id__merge_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergeProposalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Merge Proposal"
      }
    },
...
  }
}